ENH: improve cspline1d, qspline1d, and relative funcs via Pythran
