+++ scipy/signal/__init__.py
@@ -299,6 +299,7 @@ from ._upfirdn import upfirdn
 from .spline import *
 
 from .bsplines import *
+from .morebsplines import *
 from .filter_design import *
 from .fir_filter_design import *
 from .ltisys import *
